third_party/blink/web_tests/virtual/mathml-disabled/sanitizer-api-math-namespace-designator.html

<!DOCTYPE html>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script>
  const source = `<math><mtext>Hello</mtext></math>`;
  test(function() {
    const sanitizer = new Sanitizer({allowElements: ["math:math", "math:mtext"]});
    assert_equals(sanitizer.sanitizeFor("template", source).innerHTML, source);
  }, `The 'math:' namespace designator allows to match MathML elements.`);
  test(function() {
    const sanitizer = new Sanitizer({allowElements: ["math", "mtext"]});
    assert_equals(sanitizer.sanitizeFor("template", source).innerHTML, "Hello");
  }, `The 'math:' namespace designator is required to match MathML elements.`);
</script>
